Make Clipper properly create polygons with holes
The original project maintainers were interested in triangulating the
areas _inside_ the holes. This commit changes the default behaviour to
triangulate _inside_ the outer polygon boundary _without_ the holes.

Make use of the ptSubject and ptClip variables for clipper to discern
between the outer polygon boundary and holes. The resulting polygon is
the result of XOR operation between the boundary and holes.